requirements.txt
annotated-types==0.7.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.181.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.74.0
grpcio-status==1.71.2
httplib2==0.31.0
idna==3.10
markdown-it-py==4.0.0
mdurl==0.1.2
orjson==3.11.3
playwright==1.55.0
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.8
pydantic_core==2.33.2
pyee==13.0.0
Pygments==2.19.2
pyparsing==3.2.4
python-dotenv==1.0.1
requests==2.32.5
rich==14.1.0
rsa==4.9.1
shellingham==1.5.4
tqdm==4.67.1
typer==0.17.4
typing-inspection==0.4.1
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0

# Testing dependencies
pytest==8.3.2
pytest-asyncio==0.25.0
pytest-cov==6.1.0
pytest-mock==3.14.0
